address minor review comments.
